+++ Sources/Sentry/PrivateSentrySDKOnly.m
@@ -284,7 +284,7 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
 #if SENTRY_TARGET_REPLAY_SUPPORTED
     // As the options are not passed in by the hybrid SDK, we need to use the options from the
     // current hub.
-    SentryScreenshotOptions *_Nonnull options = SentrySDK.options.screenshot;
+    SentryViewScreenshotOptions *_Nonnull options = SentrySDK.options.screenshot;
     SentryViewScreenshotProvider *_Nonnull screenshotProvider =
         [SentryDependencyContainer.sharedInstance getScreenshotProviderForOptions:options];
     return [screenshotProvider appScreenshotsData];