+++ Sources/Swift/Core/Tools/ViewCapture/SentryViewScreenshotProvider.swift
@@ -8,6 +8,7 @@ import UIKit
 @objc
 @_spi(Private) public protocol SentryViewScreenshotProvider: NSObjectProtocol {
     func image(view: UIView, onComplete: @escaping ScreenshotCallback)
+    func appScreenshots() -> [UIImage]
 }
 #endif
 #endif